+++ mobile/android/fenix/app/src/androidTest/java/org/mozilla/fenix/ui/robots/ThreeDotMenuMainRobot.kt
@@ -27,6 +27,7 @@ import androidx.test.uiautomator.Until
 import org.hamcrest.Matchers.allOf
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
+import org.mozilla.fenix.FeatureFlags
 import org.mozilla.fenix.R
 import org.mozilla.fenix.helpers.Constants.LONG_CLICK_DURATION
 import org.mozilla.fenix.helpers.Constants.RETRY_COUNT
@@ -101,10 +102,17 @@ class ThreeDotMenuMainRobot {
             addToHomeScreenButton,
             addToShortcutsButton,
             saveToCollectionButton,
-            settingsButton(),
         )
         assertCheckedItemWithResIdAndTextExists(addBookmarkButton)
         assertCheckedItemWithResIdAndTextExists(desktopSiteToggle(isRequestDesktopSiteEnabled))
+        // Swipe to second part of menu
+        expandMenu()
+        assertItemContainingTextExists(
+            settingsButton(),
+        )
+        if (FeatureFlags.print) {
+            assertItemContainingTextExists(printContentButton)
+        }
         assertItemWithDescriptionExists(
             backButton,
             forwardButton,
@@ -582,6 +590,7 @@ private val reportSiteIssueButton = itemContainingText("Report Site Issue")
 private val addToHomeScreenButton = itemContainingText(getStringResource(R.string.browser_menu_add_to_homescreen))
 private val addToShortcutsButton = itemContainingText(getStringResource(R.string.browser_menu_add_to_shortcuts))
 private val saveToCollectionButton = itemContainingText(getStringResource(R.string.browser_menu_save_to_collection_2))
+private val printContentButton = itemContainingText(getStringResource(R.string.menu_print))
 private val backButton = itemWithDescription(getStringResource(R.string.browser_menu_back))
 private val forwardButton = itemWithDescription(getStringResource(R.string.browser_menu_forward))
 private val shareButton = itemWithDescription(getStringResource(R.string.share_button_content_description))